documents list upload now in same component, documents can be deleted.
