+++ src/App.tsx
@@ -3,7 +3,7 @@ import { Route, Routes, useNavigate } from 'react-router-dom';
 import './App.css';
 import { AddStopword } from './components/AddStopword';
 import { DocumentListPage } from './components/DocumentListPage';
-import { FileUploadPage } from './components/FileUploadPage';
+import { HomePage } from './components/HomePage';
 import { StopWordsPage } from './components/StopWordsPage';
 
 const { Header, Content } = Layout;
@@ -11,7 +11,7 @@ const { Header, Content } = Layout;
 const menuItems: MenuProps['items'] = [
   {
     key: "/",
-    label: "Dokumentum feltöltés",
+    label: "Home",
     style: { width: "25%" }
   },
   {
@@ -53,7 +53,7 @@ function App() {
         </Header>
         <Content style={{ padding: '0 50px' }}>
           <Routes>
-            <Route path='/' element={<FileUploadPage />} />
+            <Route path='/' element={<HomePage />} />
             <Route path='/documents' element={<DocumentListPage />} />
             <Route path='/stopwords' element={<StopWordsPage />} />
             <Route path='/stopwords/add' element={<AddStopword />} />
