+++ src/components/HomePage.tsx
@@ -1,0 +1,7 @@
+export function HomePage() {
+    return (
+        <div className="home_page">
+            Parsec applikáció.
+        </div>
+    )
+}
